requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0
statsmodels>=0.13.0
matplotlib>=3.4.0
seaborn>=0.11.0
h5py>=3.1.0
umap-learn>=0.5.0
leidenalg>=0.8.0
louvain>=0.7.0
igraph>=0.9.0